gwproxy: Migrate epoll specific code to epoll.c
The main migration of epoll specific code to epoll.c. It's in
preparation to add io_uring support.
